refactor(extension): modularize webview and settings view internals
Split oversized webview/settings frontend components and the settings
message handler along concern seams behind unchanged entry points;
modernize type-only imports and narrowing.
